Add support for Django 1.5's streaming response
Well, by "add support" I mean provide a 'streaming' attribute to the
response class that allows Django's testrunner to not choke when
processing the response.  Without this, you get an AttributeError from here:
https://github.com/django/django/blob/stable/1.5.x/django/test/testcases.py#L639

I think the right fix is in django-webtest rather than in Django itself.

The commit ID in tox.ini for Django 1.5 was 4 months out of date.  When
I bumped it to today's latest commit, the tests started failing (hence I
didn't need to write a new failing test).