Don't add duplicate WebtestUserMiddleware to the list of middlewares in WebTestMixin
I ran into problems when adding a large number of tests to our repo that all used django_app_factory. Every test adds a new instance of WebtestUserMiddleware to the WebTestMixin.settings_middleware, which eventually leads to hitting the max stack size and causing tests to fail.

In WebTestMixin._setup_auth_middleware, added a check for whether WebTestMixin is already in the list in before adding it.