Fix `WebtestUserMiddleware` that no longer inherits `MiddlewareMixin`
In the following Django PR targeting 5.2, the Django `RemoteUserMiddleware` class stopped inheriting from `MiddlewareMixin`. This meant that the `WebtestUserMiddleware` stopped being called, even if it was installed, as it did not implement the necessary `__call__` method.

https://github.com/django/django/pull/18036/files#diff-0855f8c0e8b7fbb0ed4cd47469c9b75c9e8f3de04fd0415d92e05b7e682fbb1e

This change fixes `WebtestUserMiddleware` to use the modern approach to middleware that was introduced in Django 1.10 and is now used by its superclass.